requirements.txt
streamlit==1.43.2
PyPDF2==3.0.1
langchain==0.3.20
langchain_google_genai==2.1.0
faiss-cpu==1.10.0
pydantic==2.10.6
numpy==2.2.3
pandas==2.2.3
protobuf==5.29.3
google-api-core==2.24.2
google-auth==2.38.0
google-ai-generativelanguage==0.6.16
tenacity==9.0.0
requests==2.32.3
urllib3==2.3.0
PyYAML==6.0.2
tornado==6.4.2
python-dotenv==1.0.1